+++ packages/core/src/annotations/PopupWrapper.tsx
@@ -18,6 +18,7 @@ interface PopupWrapperProps {
 }
 
 const PopupWrapper: React.FC<PopupWrapperProps> = ({ annotation }) => {
+    const containerRef = React.createRef<HTMLDivElement>();
     const theme = React.useContext(ThemeContext);
     let dateStr = '';
     if (annotation.modificationDate) {
@@ -25,8 +26,32 @@ const PopupWrapper: React.FC<PopupWrapperProps> = ({ annotation }) => {
         dateStr = date ? `${date.toLocaleDateString()}, ${date.toLocaleTimeString()}` : '';
     }
 
+    React.useLayoutEffect(() => {
+        // Increase the z-index of annotation element
+        // so it won't be displayed under the previous or next page
+        const containerEle = containerRef.current;
+        if (!containerEle) {
+            return;
+        }
+
+        // Query the annotation element
+        const annotationEle = document.querySelector(`[data-annotation-id="${annotation.id}"]`);
+        if (!annotationEle) {
+            return;
+        }
+
+        const ele = (annotationEle as HTMLElement);
+        ele.style.zIndex += 1;
+
+        return () => {
+            // Reset the `z-index` when the popup is closed
+            ele.style.zIndex = `${parseInt(ele.style.zIndex, 10) - 1}`;
+        };
+    }, []);
+
     return (
         <div
+            ref={containerRef}
             className={`${theme.prefixClass}-annotation-popup-wrapper`}
             style={{
                 top: annotation.annotationType === AnnotationType.Popup ? '' : '100%',